+++ tests/vaitk/core/drivers/text/test_text_screen_driver.py
@@ -1,0 +1,2 @@
+def test_instantiation():
+    pass
